+++ R/offline.R
@@ -144,14 +144,27 @@ reconstructGPlates <- function(x, age, model, path.gplates=NULL,dir=NULL, verbos
 		# folder where files will be executed
 		if(is.null(dir)) tempd <- tempdir() else tempd <- dir
 
-		# copy all model feature to working directory
+		# copy all model feature to working directory - unless it is a shapefile
 		sources <- plateFeatures
-		plateFeatures <-file.path(tempd, fileFromPath(sources, win=win)) 
-		# bug fix for windows
-		if (win) plateFeatures <- gsub("/","\\\\", plateFeatures)
 
+		for(i in 1:length(sources)){
+
+			# if it is not a shapefile, copy it over
+			if(!isShapefile(sources[i])){
+				# a single
+				plateFeature <-file.path(tempd, fileFromPath(sources[i], win=win))
+
+				# bug fix for windows
+				if (win) plateFeature <- gsub("/","\\\\", plateFeature)
+
+				# copy it over
+				results <- file.copy(sources[i], plateFeature)
+			}else{
+				plateFeature <- sources[i]
+			}
+			plateFeatures[i] <- plateFeature
+		}
 		names(plateFeatures) <- names(sources)
-		results <- file.copy(sources, plateFeatures)
 
 		# prepare x
 		# create a SpatialPointsDataFrame from long-lat matrix
@@ -173,6 +186,7 @@ reconstructGPlates <- function(x, age, model, path.gplates=NULL,dir=NULL, verbos
 
 			# transform to sf
 			x <- sf::st_as_sf(x, coords=c("lng", "lat"))
+			
 		}
 		
 	
@@ -221,6 +235,13 @@ reconstructGPlates <- function(x, age, model, path.gplates=NULL,dir=NULL, verbos
 			# if there: select the partitioning feature collection
 			platePolygons <- plateFeatures[partitioning]
 
+			# and then check if it a shapefile
+			if(isShapefile(platePolygons)){
+				# if it is, then it must be converted to a gpml
+				platePolygons <- shp_to_gpml(platePolygons,  dir=tempd,
+					gplatesExecutable=gplatesExecutable, winin=win, winout=win, verbose=verbose)
+			}
+
 
 		}else{
 		# feature to reconstruct is the static polygons
@@ -228,6 +249,13 @@ reconstructGPlates <- function(x, age, model, path.gplates=NULL,dir=NULL, verbos
 			# look for x in the feature set
 			validFeature <- any(x==names(plateFeatures))
 			if(validFeature){
+
+				# if it a shapefile, then it needs to be converted to a gpml
+				if(isShapefile(plateFeatures[x])){
+					# convert here
+					plateFeatures[x] <- shp_to_gpml(plateFeatures[x],  dir=tempd,
+						gplatesExecutable=gplatesExecutable, winin=win, winout=win, verbose=verbose)
+				}
 
 				# which file version is used? .gmpl or .gpmlz?
 				withZ <- grepl(".gpmlz", plateFeatures[x])
@@ -280,6 +308,9 @@ reconstructGPlates <- function(x, age, model, path.gplates=NULL,dir=NULL, verbos
 		# the single file
 		targetSingle <- paste(pathToFileNoEXT,"_reconstructed.shx",	sep="")
 		targetSingleNoPath <- fileFromPath(targetSingle, win=win)
+
+		if(verbose) message(paste0("Reading in: ", targetSingle))
+
 
 		# produced directory? 
 		targetDir<- paste(pathToFileNoEXT,"_reconstructed",	sep="")
@@ -523,5 +554,62 @@ mergeRotations <- function(x){
 
 	# return the name of the new rotation file
 	return(newFile)
+
+}
+
+
+isShapefile <- function(x){
+	# get filename
+	filename <- rgplates:::fileFromPath(x)
+	grepl("\\.shx$",filename) | grepl("\\.shp$",filename)
+}
+
+shp_to_gpml <- function(x, dir=file.path(tempdir(), "newgpml"), gplatesExecutable, winin=FALSE, winout=FALSE, verbose=FALSE){
+	if(verbose){
+		message("Transforming shapefile to gpml:")
+		message(x)
+	}
+
+	# first ensure that the directory exists
+	dir.create(dir, showWarnings=FALSE)
+
+	if(winin) x <- gsub("\\\\", "/", x)
+
+	# copy over all relevant files
+	filename <- rgplates:::fileFromPath(x)
+
+	# the filename stem
+	fileStem <- unlist(lapply(strsplit(filename, "\\."), function(x) paste(x[-length(x)], collapse="")))
+
+	# origdir
+	origdir <- unlist(lapply(strsplit(x, "/"), function(x) paste(x[-length(x)], collapse="/")))
+
+	# list out all the files (shapefiles should have the same stem
+	allFiles <- list.files(origdir)
+
+
+	# copy the relevant files
+	toCopy <- allFiles[grep(fileStem, allFiles)]
+
+
+	for(i in 1:length(toCopy)){
+		file.copy(file.path(origdir,toCopy[i]),file.path(dir, toCopy[i]))
+	}
+
+	# look for the shapefile (shp)
+	shpFile <- toCopy[grep("\\.shp$",toCopy)]
+	pathToFile <- file.path(dir, shpFile)
+
+	# now check the OS
+	if (winout) pathToFile <- gsub("/","\\\\", pathToFile)
+
+	# then use gplates
+	conversion <- paste(gplatesExecutable, " convert-file-format -l \"",pathToFile,"\" -e gpml",sep="")
+
+	# execute conversion
+	system(conversion, ignore.stdout=!verbose,ignore.stderr=!verbose)
+
+	# return path to gpml file
+	return(paste0(dir, "/", fileStem, ".gpml"))
 
 }