+++ NEWS.md
@@ -11,7 +11,9 @@
 - The `sawteeth` function as the engine for visualizing subduction lines.
 - Plot method for the `subduction.zones` S3 class (`plot.subduction.zones`)
 - Support for `.gpmlz` feature collection files. 
-- Support for multiple rotation files per tectonic model. Muliple files will be concatenated at recontruction time. 
+- Support for multiple rotation files per tectonic model. The rotation files will be concatenated at recontruction time. 
+- Support for shapefile model feature collections.
+- `platemodel` constructor enforces the presence of linked files.
 
 ### Changed
 
@@ -20,6 +22,7 @@
 ### Fixed
 
 - In the offline `reconstruct` method, if the `partitioning` feature collection is not present, the function will indicate it with a correct error message. 
+
 
 # rgplates 0.5.0 - 2024-09-04